🔑 Authorization complete

+++ backend/routes/posts.js
@@ -47,6 +47,7 @@ router.post("", checkAuth, multer({storage: storage}).single("image") ,(req,res,
     title: req.body.title,
     content: req.body.content,
     imagePath: url+"/images/"+req.file.filename,
+    creator: req.userData.userId
   });
   // console.log(post);
   post.save().then(result => {
@@ -58,7 +59,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image") ,(req,res,
         id: result._id,
         title: result.title,
         content: result.content,
-        imagePath: result.imagePath
+        imagePath: result.imagePath,
+        creator: result.creator
       }
     });
   });
@@ -99,21 +101,32 @@ multer({storage: storage}).single("image"),
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  console.log(post);
-  Post.updateOne({_id:req.params.id},post).then((result)=>{
-    console.log(result);
-    res.status(200).json({message: "update successfull"});
+  Post.updateOne({_id:req.params.id, creator:req.userData.userId },post).then(
+    result=>{
+      console.log(result);
+      if(result.nModified > 0){
+        res.status(200).json({message: "update successfull"});
+      } else {
+        res.status(401).json({message:"unauthorized!"});
+      }
   });
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id:req.params.id}).then(result =>{
+  Post.deleteOne({_id:req.params.id, creator:req.userData.userId}).then(result =>{
     console.log(result);
-  });
-  res.status(200).json({
-    message: "Post deleted"
+    if(result.n>0){
+      res.status(200).json({
+        message: "Post deleted"
+      });
+    } else {
+      res.status(401).json({
+        message: "unauthorized"
+      });
+    }
   });
 });
